refactor(state): extract spawn deletion aggregate module
Move published-spawn deletion and its lock-order composition into
spawn_aggregate.py so spawn_store stays under the size guardrail.
