+++ src/meridian/lib/state/spawn_store.py
@@ -8,8 +8,7 @@ from __future__ import annotations
 import os
 import secrets
 import shutil
-import stat
-from collections.abc import Callable, Mapping
+from collections.abc import Mapping
 from contextlib import suppress
 from datetime import UTC, datetime
 from pathlib import Path
@@ -38,7 +37,6 @@ from meridian.lib.core.types import SpawnId
 from meridian.lib.state.atomic import atomic_publish_dir, atomic_write_text
 from meridian.lib.state.event_store import lock_file
 from meridian.lib.state.paths import RuntimePaths
-from meridian.lib.state.process_scope_projection import scope_projection_lock_path
 from meridian.lib.state.spawn.model import (
     CancelIntent as CancelIntent,
 )
@@ -58,7 +56,7 @@ from meridian.lib.state.spawn.repository import (
     is_safe_spawn_dir_name as _is_safe_spawn_dir_name,
 )
 from meridian.lib.state.spawn.repository import read_state as _read_state
-from meridian.lib.state.spawn.repository import record_to_stored_state, spawn_lock_path
+from meridian.lib.state.spawn.repository import record_to_stored_state
 from meridian.lib.state.spawn.repository import scan_spawn_ids as _scan_spawn_ids
 from meridian.lib.state.spawn.repository import write_state_locked as _write_state_locked
 from meridian.lib.state.spawn.terminal_policy import decide_terminal_write
@@ -70,10 +68,9 @@ from meridian.lib.state.spawn.transitions import (
     apply_record_exited,
     apply_runner_exit,
 )
+from meridian.lib.state.spawn_aggregate import delete_published_spawn
 
 logger = structlog.get_logger(__name__)
-
-type SpawnDeletionPrecondition = Callable[[SpawnRecord | None], bool]
 
 # ---------------------------------------------------------------------------
 # ID generation (absorbed from state/id_gen.py)
@@ -243,61 +240,6 @@ def gc_abandoned_stages(runtime_root: Path) -> None:
                     shutil.rmtree(entry)
                 else:
                     entry.unlink()
-
-
-def _restore_spawn_artifact_permissions(
-    func: Callable[[str], object],
-    path: str,
-    exc_info: BaseException,
-) -> None:
-    if isinstance(exc_info, FileNotFoundError):
-        return
-    with suppress(OSError):
-        os.chmod(path, stat.S_IWRITE)
-    try:
-        func(path)
-    except OSError as error:
-        raise exc_info from error
-
-
-def delete_published_spawn(
-    runtime_root: Path,
-    spawn_id: SpawnId | str,
-    *,
-    can_delete: SpawnDeletionPrecondition,
-) -> bool:
-    """Delete one published spawn when its locked aggregate permits it.
-
-    This aggregate seam composes the spawn-state and process-scope persistence
-    leaves. Every published-row deletion routes through it. A cleanup claim
-    prevents deletion because it is durable at-least-once intent: the reaper
-    must finish or clear the claim before artifact retention may remove it.
-    Callers that also need ``spawns_flock`` must acquire it first.
-    """
-
-    paths = RuntimePaths.from_root_dir(runtime_root)
-    resolved_spawn_id = str(spawn_id)
-    if not _is_safe_spawn_dir_name(resolved_spawn_id):
-        raise ValueError(f"Invalid spawn ID: {resolved_spawn_id}")
-    spawn_dir = paths.spawns_dir / resolved_spawn_id
-
-    # Global order: spawn state, then process-scope projection.
-    with (
-        lock_file(spawn_lock_path(paths.spawns_dir, resolved_spawn_id)),
-        lock_file(scope_projection_lock_path(runtime_root, resolved_spawn_id)),
-    ):
-        claim_path = spawn_dir / "reaper_cleanup_claim.json"
-        if claim_path.exists() or not can_delete(
-            _read_state(paths.spawns_dir, resolved_spawn_id, include_prompt=False)
-        ):
-            return False
-        if not spawn_dir.exists():
-            return False
-        try:
-            shutil.rmtree(spawn_dir, onexc=_restore_spawn_artifact_permissions)
-        except OSError:
-            return False
-        return True
 
 
 def start_spawn(
